examples/magickey_auth/requirements.txt
plash_cli @ git+https://github.com/AnswerDotAI/plash_cli@magickey
python-fasthtml @ git+https://github.com/AnswerDotAI/fasthtml@magickey
webauthn
numpy